// frontend/src/app/core/models/enums.ts
export enum Gender {
  Men = 'MEN',
  Women = 'WOMEN'
}

export enum SubCategory {
  TShirt = 'T_SHIRT',
  Pants = 'PANTS',
  Shirt = 'SHIRT',
  Jeans = 'JEANS',
  Hoodie = 'HOODIE',
  Jacket = 'JACKET',
  Dress = 'DRESS',
  Skirt = 'SKIRT'
}

export enum OrderStatus {
  Pending = 0,
  Confirmed = 1,
  Shipped = 2,
  Delivered = 3,
  Cancelled = 4,
  Failed = 5
}

export enum UserRole {
  Customer = 'CUSTOMER',
  Admin = 'ADMIN'
}

export enum AppLanguage {
  Arabic = 'ar',
  English = 'en'
}